Use node to run TS scripts instead of jiti
Avoid needing to download a dependency to run a script, which is slow,
and can trigger minimumReleaseAge errors, as it always downloads the
newest version. Pinning the download to a version is possible, but ends
up repeating version info since we also need jiti for ESLint.

This works because we require node 24, which supports running TS scripts
without the need for compilation.